+++ test/android_id_test.dart
@@ -16,11 +16,38 @@ void main() {
   });
 
   tearDown(() {
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
-        .setMockMethodCallHandler(channel, null);
+    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger.setMockMethodCallHandler(channel, null);
   });
 
   test('getAndroidId', () async {
+    expect(await plugin.getId(), '42');
+  });
+
+  test('returns null on non-Android platforms', () async {
+    addTearDown(() => debugDefaultTargetPlatformOverride = null);
+
+    debugDefaultTargetPlatformOverride = TargetPlatform.fuchsia;
+
+    expect(await plugin.getId(), isNull);
+
+    debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
+
+    expect(await plugin.getId(), isNull);
+
+    debugDefaultTargetPlatformOverride = TargetPlatform.linux;
+
+    expect(await plugin.getId(), isNull);
+
+    debugDefaultTargetPlatformOverride = TargetPlatform.macOS;
+
+    expect(await plugin.getId(), isNull);
+
+    debugDefaultTargetPlatformOverride = TargetPlatform.windows;
+
+    expect(await plugin.getId(), isNull);
+
+    debugDefaultTargetPlatformOverride = TargetPlatform.android;
+
     expect(await plugin.getId(), '42');
   });
 }